[AWARE] Fix update publish/subscribe
When updating an existing publish/subscribe discovery session the
native API must take the session ID of the existing session. The code
provided 0 - which was an indication to create a new discovery session.

Bug: 38397363
Test: unit tests and integration (sl4a) tests

+++ service/java/com/android/server/wifi/aware/WifiAwareDiscoverySessionState.java
@@ -42,14 +42,14 @@ public class WifiAwareDiscoverySessionState {
 
     private final WifiAwareNativeApi mWifiAwareNativeApi;
     private int mSessionId;
-    private int mPubSubId;
+    private byte mPubSubId;
     private IWifiAwareDiscoverySessionCallback mCallback;
     private boolean mIsPublishSession;
 
     private final SparseArray<String> mMacByRequestorInstanceId = new SparseArray<>();
 
     public WifiAwareDiscoverySessionState(WifiAwareNativeApi wifiAwareNativeApi, int sessionId,
-            int pubSubId, IWifiAwareDiscoverySessionCallback callback, boolean isPublishSession) {
+            byte pubSubId, IWifiAwareDiscoverySessionCallback callback, boolean isPublishSession) {
         mWifiAwareNativeApi = wifiAwareNativeApi;
         mSessionId = sessionId;
         mPubSubId = pubSubId;
